Add with-environment and with-write-transaction to lmdb-utils
Attempting to open multiple LMDB write transactions from the same
process causes problems, so with-write-transaction introduces a mutex
so srfi-18 threads can wait until a write transaction become
available. The mutex is stored in a global hash-table keyed by LMDB
environment handle.

with-environment is a more convenient way to open/close an environment
and clean up afterwards (in particular, removing entries in the global
hash table used by with-write-transaction).